--- src/test/java/nextstep/acceptance/AcceptanceTest.java
@@ -1,19 +1,0 @@
-package nextstep.acceptance;
-
-import nextstep.utils.DatabaseCleanup;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@ActiveProfiles("test")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class AcceptanceTest {
-    @Autowired
-    private DatabaseCleanup databaseCleanup;
-
-    @BeforeEach
-    void setUp() {
-        databaseCleanup.execute();
-    }
-}
