+++ src/test/java/nextstep/station/acceptance/StationApiRequest.java
@@ -1,4 +1,4 @@
-package nextstep.acceptance;
+package nextstep.station.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
